More debugging

+++ GameEngineBase/src/Platform/Win32/Win32InputManager.cpp
@@ -31,7 +31,7 @@ namespace gebase {
 
 	void InputManager::setMousePosition(const math::Vector2f& pos)
 	{
-		POINT p = { (LONG)pos.x, (LONG)pos.y };
+		POINT p = { pos.x, pos.y };
 		ClientToScreen(hWnd, &p);
 		SetCursorPos(p.x, p.y);
 	}
